landing page now uses a customised hero widgets
the new widget accepts two images rather than one
which are superimposed on one another. the idea is
to be able to superimpose a webp bitmap on top of
an svg artwork while also being able to position
them on top of a page-wide gradient which is visible
in the landing page after this commit.